Implement matmul_simd_j
This uses GCC's built-in short-vector implementation, in the hope that
it might produce more stable code.  This seems to work -- I can get it
to actually emit the code I would have expected the pvec
implementation to emit.

I also now compile on clang, gcc-4.8, and gcc-4.9 -- the idea here is
that they might behave a bit differently.  They certianly seem to!